+++ NEWS.md
@@ -14,6 +14,8 @@
 
 * The `by` parameter in `$set_*()` methods, deprecated in 0.2.0, is now defunct.
 
+* The `$set_equal_contributor()` method, deprecated in 0.2.0, is now defunct.
+
 # plume 0.2.5
 
 * `$get_author_list(suffix =)` is now slightly more flexible and will try less hard to sanitise author list suffixes (#90).